Fixed conflict resolution
