+++ KeyVault.Acmebot/Internal/CertificatePolicyExtensions.cs
@@ -1,6 +1,4 @@
-﻿using System.Linq;
-
-using Azure.Security.KeyVault.Certificates;
+﻿using Azure.Security.KeyVault.Certificates;
 
 using KeyVault.Acmebot.Models;
 
@@ -8,21 +6,6 @@ namespace KeyVault.Acmebot.Internal;
 
 internal static class CertificatePolicyExtensions
 {
-    public static CertificatePolicyItem ToCertificatePolicyItem(this CertificatePolicy certificatePolicy, string certificateName)
-    {
-        var dnsNames = certificatePolicy.SubjectAlternativeNames.DnsNames.ToArray();
-
-        return new CertificatePolicyItem
-        {
-            CertificateName = certificateName,
-            DnsNames = dnsNames.Length > 0 ? dnsNames : new[] { certificatePolicy.Subject[3..] },
-            KeyType = certificatePolicy.KeyType?.ToString(),
-            KeySize = certificatePolicy.KeySize,
-            KeyCurveName = certificatePolicy.KeyCurveName?.ToString(),
-            ReuseKey = certificatePolicy.ReuseKey
-        };
-    }
-
     public static CertificatePolicy ToCertificatePolicy(this CertificatePolicyItem certificatePolicyItem)
     {
         var subjectAlternativeNames = new SubjectAlternativeNames();
