+++ KeyVault.Acmebot/Internal/DnsZoneExtensions.cs
@@ -1,8 +1,8 @@
-﻿using KeyVault.Acmebot.Models;
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 
+using KeyVault.Acmebot.Models;
 using KeyVault.Acmebot.Providers;
 
 namespace KeyVault.Acmebot.Internal;
@@ -13,7 +13,7 @@ internal static class DnsZoneExtensions
     {
         return new DnsZoneItem { Name = dnsZone.Name, DnsProviderName = dnsZone.DnsProvider.Name };
     }
-    
+
     public static DnsZone FindDnsZone(this IEnumerable<DnsZone> dnsZones, string dnsName)
     {
         return dnsZones.Where(x => string.Equals(dnsName, x.Name, StringComparison.OrdinalIgnoreCase) || dnsName.EndsWith($".{x.Name}", StringComparison.OrdinalIgnoreCase))
